// patterns/pattern-typings/AuxLibraryDefinitions.d.ts
declare type EaseType =
  | "linear"
  | "quadratic"
  | "cubic"
  | "quartic"
  | "quintic"
  | "sinusoidal"
  | "exponential"
  | "circular"
  | "elastic";

declare type EaseMode = "in" | "out" | "inout";

declare interface Easing {
  type: EaseType;
  mode: EaseMode;
}